# src/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "",
        "version": "1.0.0"
    },
    "paths": {
        "/api/auth/register": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Register",
                "description": "Register user by either email/phone",
                "operationId": "authRegister",
                "requestBody": {
                    "description": "Pass user credentials",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "phone"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "user@mail.com"
                                    },
                                    "phone": {
                                        "type": "integer",
                                        "format": "string",
                                        "example": "+923215508055"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success or success with email verification sent",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "": {
                                            "type": "object",
                                            "example": {
                                                "saved": true,
                                                "message": "user->email ? A key has been sent to your email to verify : User created successfully"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Wrong input response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "The given data was invalid."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/auth/register-info": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Register user info",
                "description": "Register user info for name, username, password",
                "operationId": "authRegisterInfo",
                "requestBody": {
                    "description": "Pass user credentials, for param user send either phone or email of user",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "username",
                                    "password",
                                    "user"
                                ],
                                "properties": {
                                    "user": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "test@kilogram.ga"
                                    },
                                    "username": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "hasnatsafder"
                                    },
                                    "name": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "Hasnat Safder"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "123456"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success or success with email verification sent",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "": {
                                            "type": "object",
                                            "example": {
                                                "saved": true,
                                                "message": "Data added successfully"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "User not found for this email",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "object",
                                            "example": {
                                                "success": false,
                                                "message": "User not found"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/auth/social-login/facebook": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Social login",
                "description": "Register user from facebook token",
                "operationId": "authSocialLogin",
                "requestBody": {
                    "description": "Pass facebook token",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "token"
                                ],
                                "properties": {
                                    "token": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "EAAPTe5h66KgBAMBC34mi9fZCgVH5SLccUM15TC4EXHZBQhqHgVkTha13WLOOgdHFV1IjwEQLZBvDJDDYSnztdBtoyYv6akisfB1rB4EY6LtcWZBFLomOaNM9xQWDaI6DHwfCvQaPiK41Yz7GNJhubAIKYM68ZAR88mJO0D2ZBCZCDxkdOpPISbcSLI4ZBHTbGYQ3AaM5U8t6xYENTzfUSzlCWBz8IlbUb5EZD"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successfully added social user",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "": {
                                            "type": "object",
                                            "example": {
                                                "access_token": "xyz",
                                                "token_type": "Bearer",
                                                "refresh_token": null,
                                                "expires_at": "2020-10-09 11:42:46"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "User not found for this token",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "object",
                                            "example": {
                                                "success": false,
                                                "message": "User not found for this token"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/auth/verify-code": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Verify code",
                "description": "Verify email code",
                "operationId": "authVerifyCode",
                "requestBody": {
                    "description": "Pass email and Verification key",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "verification_key"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "test@kilogram.ga"
                                    },
                                    "verification_key": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "985654"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Verification successful",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "": {
                                            "type": "object",
                                            "example": {
                                                "success": true,
                                                "message": "Verification successful"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "User not found for this token",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "object",
                                            "example": {
                                                "success": false,
                                                "message": "Verification code not found"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/auth/validate_token/{token}": {
            "get": {
                "tags": [
                    "Auth"
                ],
                "summary": "Validate token",
                "description": "Validate a reset token",
                "operationId": "authValidateToken",
                "parameters": [
                    {
                        "name": "token",
                        "in": "path",
                        "description": "Token",
                        "required": true,
                        "example": "1nOsTmJM4Ne7d8cqyG5aeJAPh49oSC7cj3OYmiC4fyCQhpxg3TEwD6ZJnmji"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Valid token",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "": {
                                            "type": "object",
                                            "example": {
                                                "success": true,
                                                "message": "Valid token"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Token not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "object",
                                            "example": {
                                                "success": false,
                                                "message": "Token not found"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
...
    "security": [
        []
    ]
}